moved up the getLastClosedAction function to handle archived room reasons
