+++ src/pages/home/ReportScreen.js
@@ -195,6 +195,12 @@ class ReportScreen extends React.Component {
             return null;
         }
 
+        const isArchivedRoom = ReportUtils.isArchivedRoom(this.props.report);
+        let reportClosedAction = null;
+        if (isArchivedRoom) {
+            reportClosedAction = ReportActionsUtils.getLastClosedReportAction(this.props.reportActions);
+        }
+
         // We are either adding a workspace room, or we're creating a chat, it isn't possible for both of these to be pending, or to have errors for the same report at the same time, so
         // simply looking up the first truthy value for each case will get the relevant property if it's set.
         const reportID = getReportID(this.props.route);
@@ -293,6 +299,8 @@ class ReportScreen extends React.Component {
                                         report={this.props.report}
                                         isComposerFullSize={this.props.isComposerFullSize}
                                         onSubmitComment={this.onSubmitComment}
+                                        isArchivedRoom={isArchivedRoom}
+                                        reportClosedAction={reportClosedAction}
                                     />
                                 </>
                             )}
